requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
sqlalchemy==2.0.41
aiosqlite==0.21.0
alembic==1.16.4
redis==5.2.1
celery==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
httpx==0.28.1
pytest==8.4.1
pytest-asyncio==1.1.0
pandas==2.2.3
cryptography==45.0.6
MetaTrader5==5.0.45; platform_system == "Windows"

